Attenuates the HF in hybrid mode to match what SILK does below the cutoff
Conflicts:

	src/opus_multistream.c
	src/opus_private.h

+++ src/opus_decoder.c
@@ -693,7 +693,7 @@ static int opus_packet_parse_impl(const unsigned char *data, opus_int32 len,
       *out_toc = toc;
 
    if (payload_offset)
-      *payload_offset = data-data0;
+      *payload_offset = (int)(data-data0);
 
    return count;
 }
